% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score_fishing.R
\name{Score_fishing}
\alias{Score_fishing}
\title{Detect fishing effort by image analysis}
\usage{
Score_fishing(x, y, map)
}
\arguments{
\item{map}{An image of class cimg.}

\item{X}{A vector of X indices for pixels OUTISDE the target area.}

\item{Y}{A vector of Y indices for pixels OUTSIDE the target area.}
}
\value{
The proportion of pixels within a region of interest which have fishing activity.
}
\description{
This function takes a collection of pixels and an image, and returns the proportion of area being fished.
}
\details{
The vectors of pixels are used to create a mask which blanks out fishing in Non-Scottish waters.
The image is converted to black and white, and a threshold of 0.4 applied to remove land features while
retaining fishing activity. The proportion of bright pixels out of the pixels of interest indicates the
amount of fishing happening in the image.
}
